+++ CHANGELOG.md
@@ -2,8 +2,15 @@
 
 ## [0.3.0] 
 - Rename application inside vscode to "Aesthetics"
+  - Change the text of commands
 - Add standard color themes 
-- Add background image in default 
+  - Added "Aesthetics #1" color theme
+- Add wallpaper support
+  - Add on/off param for wallpaper
+  - Add blur intensity config parameter
+  - Add opacity config parameter
+  - Add custom wallpaper URL support
+- Additional security checks to prevent unintended injections
 
 ## [0.2.4] - 2021-11-27
 - Fix 0.2.3 issue where extension would not find any command